defining types for opengraphscrapper
