--- src/utils/expType.ts
@@ -1,12 +1,0 @@
-export interface ExperienceType {
-  id: string;
-  during: string;
-  title: string;
-  body: string;
-}
-
-export interface SkillType {
-  id: string;
-  name: string;
-  img: string;
-}
